UserEmpresas noacepta duplicados. Con clave compuesta en BBDD y en Formrequest

+++ app/Http/Controllers/UserEmpresaController.php
@@ -4,17 +4,27 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
-use App\{UserEmpresa,Empresa};
+use App \{
+    UserEmpresa,
+    Empresa,
+    User
+};
+
+use App\Http\Requests\UserEmpresaRequest;
+
+use Illuminate\Support\Facades\Validator;
+use Dotenv\Exception\ValidationException;
+
 
 class UserEmpresaController extends Controller
 {
 
     public function empAsoc($userid)
     {
-        $empAsoc=DB::table('user_empresas')
+        $empAsoc = DB::table('user_empresas')
             ->join('empresas', 'empresas.id', '=', 'user_empresas.empresa_id')
-            ->select('user_empresas.id as idUserEmp','empresa_id as idEmp','user_id as idUser','empresas.name as empresa')
-            ->where('user_id',$userid)
+            ->select('user_empresas.id as idUserEmp', 'empresa_id as idEmp', 'user_id as idUser', 'empresas.name as empresa')
+            ->where('user_id', $userid)
             ->orderBy('empresas.name')
             ->get();
 
@@ -25,8 +35,8 @@ class UserEmpresaController extends Controller
 
     public function empDisp($userid)
     {
-        $empDisp=DB::table('empresas')
-            ->select('name as empresa','id as idEmp')
+        $empDisp = DB::table('empresas')
+            ->select('name as empresa', 'id as idEmp')
             ->whereNotIn('id', function ($query) use ($userid) {
                 $query->select('empresa_id')->from('user_empresas')->where('user_id', '=', $userid);
             })
@@ -38,43 +48,52 @@ class UserEmpresaController extends Controller
         );
     }
 
-    public function store(Request $request,$userId,$empresaId)
+    public function store(UserEmpresaRequest $request)
     {
-        if ($request->ajax()) {
-            
-            $userEmpresa=DB::table('user_empresas')->insertGetId([
-                'empresa_id' => $empresaId,
-                'user_id' => $userId]
-            );
-            $empresa=DB::table('empresas')
-                 ->where('id',$empresaId)
-                 ->first();
+        DB::table('user_empresas')->insertGetId([
+            'empresa_id' => $request->empresaId,
+            'user_id' =>  $request->userId
+        ]);
 
-            $nam=$empresa->name;
+        $empresa = DB::table('empresas')
+            ->where('id', $request->empresaId)
+            ->first();
 
-            return response()->json([
-                'idUserEmp'=>$userEmpresa,
-                'nam'=>$nam,
-            ]);
-        }
+        $nam = $empresa->name;
+
+        return response()->json([
+            // $request->empresaId,
+            // $request->userId,
+            'idUserEmp' => $request->userId,
+            'nam' => $nam,
+        ]);
     }
 
     // public function destroy($userId,$idUserEmp)
-    public function destroy(Request $request,$userId,$idUserEmp,$idEmp)
+    public function destroy(Request $request)
     {
         if ($request->ajax()) {
-            $empresa=UserEmpresa::findOrFail($idUserEmp);
-            // $emp=Empresa::find($idEmp);
-            $emp=  DB::table('empresas')->where('id',$idEmp)->first();
-            $empresa->delete();
 
-            return response()->json([
-                'idEmp'=>$idEmp,
-                'idUser'=>$userId,
-                'nam'=>$emp->name,
-                'idUserEmp'=>$idUserEmp,
+            // $empresa = UserEmpresa::findOrFail($request->userEmpId);
+            $emp =  DB::table('empresas')->where('id', $request->empresaId)->first();
+            // $empresa = UserEmpresa::where('empresa_id', $request->empresa_id)->where('user_id', $request->user_id)->delete();
+            $empresa = DB::table('user_empresas')->where('empresa_id', $request->empresa_id)->where('user_id', $request->user_id)->delete();
+
+            return response()->json(
+                [
+                    'em' => $empresa,
+                    'idEmp' => $request->userEmpId,
+                    'idUser' => $request->userId,
+                    'nam' => $emp->name,
+                    'idUserEmp' => $request->userEmpId,
                 ]
-            ); 
+            );
         }
+    }
+
+    public function duplicadosUserEmpresa($userId, $empresaId)
+    {
+        $emp = UserEmpresa::where('user_id', $userId)->andWhere('empresa_id', $empresaId)->findOrFail();
+        return $emp;
     }
 }
